OTHER: Return an int in xmmsv_coll_idlist_get_size
It previously returned a size_t which was inconsistent with the rest
of the xmmsclient api.

+++ src/include/xmmsc/xmmsv_coll.h
@@ -43,7 +43,7 @@ int xmmsv_coll_idlist_remove (xmmsv_coll_t *coll, int index);
 int xmmsv_coll_idlist_clear (xmmsv_coll_t *coll);
 int xmmsv_coll_idlist_get_index (xmmsv_coll_t *coll, int index, int32_t *val);
 int xmmsv_coll_idlist_set_index (xmmsv_coll_t *coll, int index, int32_t val);
-size_t xmmsv_coll_idlist_get_size (xmmsv_coll_t *coll);
+int xmmsv_coll_idlist_get_size (xmmsv_coll_t *coll);
 
 xmmsv_coll_type_t xmmsv_coll_get_type (xmmsv_coll_t *coll);
 struct xmmsv_St *xmmsv_coll_idlist_get (xmmsv_coll_t *coll);
